# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalkYemek API
  version: dev
  description: Beylikdüzü odaklı QR + Wallet yemek sistemi
paths:
...
  /api/v1/catalog/businesses/:
    get:
      operationId: catalog_business_list
      parameters:
      - in: query
        name: district
        schema:
          enum:
          - BEYLIKDUZU
          type: string
          minLength: 1
        description: '* `BEYLIKDUZU` - Beylikdüzü'
      tags:
      - discovery
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicBusinessListResponse'
          description: ''
  /api/v1/catalog/businesses/{business_id}/:
    get:
      operationId: catalog_business_detail
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - discovery
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicBusinessDetailResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
  /api/v1/catalog/businesses/{business_id}/menu/:
    get:
      operationId: catalog_business_menu
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - discovery
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicBusinessMenuResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
...
  /api/v1/discovery/categories/:
    get:
      operationId: discovery_category_list
      parameters:
      - in: query
        name: district
        schema:
          enum:
          - BEYLIKDUZU
          type: string
          default: BEYLIKDUZU
          minLength: 1
        description: '* `BEYLIKDUZU` - Beylikdüzü'
      tags:
      - discovery
      security:
      - jwtAuth: []
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryCategoryListResponse'
          description: ''
  /api/v1/discovery/categories/{category_slug}/businesses/:
    get:
      operationId: discovery_category_business_list
      parameters:
      - in: path
        name: category_slug
        schema:
          type: string
        required: true
      - in: query
        name: district
        schema:
          enum:
          - BEYLIKDUZU
          type: string
          default: BEYLIKDUZU
          minLength: 1
        description: '* `BEYLIKDUZU` - Beylikdüzü'
      - in: query
        name: featured_first
        schema:
          type: boolean
          default: true
      - in: query
        name: listing_type
        schema:
          enum:
          - CONTRACTED
          - VOLUNTEER
          type: string
          minLength: 1
        description: |-
          * `CONTRACTED` - Contracted
          * `VOLUNTEER` - Volunteer
      - name: ordering
        required: false
        in: query
        description: Sonuçların sıralanmasında kullanılacak alan.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Sayfalanmış sonuç kümesinde bir sayfa numarası.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Her sayfada döndürülecek sonuç sayısı.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Bir arama terimi.
        schema:
          type: string
      tags:
      - discovery
      security:
      - jwtAuth: []
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDiscoveryBusinessCardList'
          description: ''
  /api/v1/discovery/home/:
    get:
      operationId: discovery_home
      parameters:
      - in: query
        name: district
        schema:
          enum:
          - BEYLIKDUZU
          type: string
          default: BEYLIKDUZU
          minLength: 1
        description: '* `BEYLIKDUZU` - Beylikdüzü'
      tags:
      - discovery
      security:
      - jwtAuth: []
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryHomeResponse'
          description: ''
...
  /api/v1/notifications/readiness/:
    get:
      operationId: notification_readiness
      tags:
      - notifications
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationReadiness'
              examples:
                NotificationReadiness:
                  value:
                    notification_ready: true
                    active_device_count: 1
                    message: ready
                  summary: Notification readiness
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
...
components:
  schemas:
...
    ApiErrorBody:
      type: object
      properties:
        code:
          type: string
        message: {}
        request_id:
          type: string
        details: {}
        reason:
          type: string
      required:
      - code
      - message
      - request_id
    ApiErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          $ref: '#/components/schemas/ApiErrorBody'
      required:
      - error
    AssetRoleEnum:
      enum:
      - GALLERY
      - COVER
      - LOGO
      - THUMBNAIL
      type: string
      description: |-
        * `GALLERY` - Gallery
        * `COVER` - Cover
        * `LOGO` - Logo
        * `THUMBNAIL` - Thumbnail
...
    DiscoveryActiveCartSummary:
      type: object
      properties:
        cart_id:
          type: integer
        business_id:
          type: integer
        business_name:
          type: string
        item_count:
          type: integer
        subtotal_amount:
          type: integer
        customer_fee_amount:
          type: integer
        total_amount:
          type: integer
      required:
      - business_id
      - business_name
      - cart_id
      - customer_fee_amount
      - item_count
      - subtotal_amount
      - total_amount
    DiscoveryBusinessCard:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        business_name:
          type: string
          maxLength: 255
        district:
          $ref: '#/components/schemas/DistrictEnum'
        district_label:
          type: string
          readOnly: true
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        listing_type_label:
          type: string
          readOnly: true
        is_featured:
          type: boolean
        short_description:
          type: string
          readOnly: true
        intro_text:
          type: string
          readOnly: true
        badge_text:
          type: string
          readOnly: true
        cover_image:
          type: string
          readOnly: true
        logo_image:
          type: string
          readOnly: true
        primary_marketplace_category:
          type: object
          additionalProperties: {}
          nullable: true
          readOnly: true
        display_priority:
          type: integer
      required:
      - badge_text
      - business_name
      - cover_image
      - district_label
      - id
      - intro_text
      - listing_type_label
      - logo_image
      - primary_marketplace_category
      - short_description
    DiscoveryCategoryListResponse:
      type: object
      properties:
        district:
          type: string
        count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryMarketplaceCategory'
      required:
      - count
      - district
      - results
    DiscoveryHomeResponse:
      type: object
      properties:
        district:
          $ref: '#/components/schemas/DistrictSummary'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryMarketplaceCategory'
        featured_businesses:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryBusinessCard'
        other_businesses:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryBusinessCard'
        active_offers:
          type: array
          items:
            $ref: '#/components/schemas/PublicOffer'
        wallet_summary:
          allOf:
          - $ref: '#/components/schemas/DiscoveryWalletSummary'
          nullable: true
        active_cart_summary:
          allOf:
          - $ref: '#/components/schemas/DiscoveryActiveCartSummary'
          nullable: true
        notification_readiness:
          $ref: '#/components/schemas/DiscoveryNotificationReadiness'
      required:
      - active_offers
      - categories
      - district
      - featured_businesses
      - notification_readiness
      - other_businesses
    DiscoveryMarketplaceCategory:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        slug:
          type: string
          maxLength: 80
          pattern: ^[-a-zA-Z0-9_]+$
        name:
          type: string
          maxLength: 120
        description:
          type: string
        sort_order:
          type: integer
        is_other:
          type: boolean
        image:
          type: string
          readOnly: true
      required:
      - id
      - image
      - name
      - slug
    DiscoveryNotificationReadiness:
      type: object
      properties:
        notification_ready:
          type: boolean
        active_device_count:
          type: integer
      required:
      - active_device_count
      - notification_ready
    DiscoveryWalletSummary:
      type: object
      properties:
        balance:
          type: integer
        pending_balance:
          type: integer
      required:
      - balance
      - pending_balance
    DistrictEnum:
      enum:
      - BEYLIKDUZU
      type: string
      description: '* `BEYLIKDUZU` - Beylikdüzü'
    DistrictSummary:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
      required:
      - code
      - label
...
    ListingTypeEnum:
      enum:
      - CONTRACTED
      - VOLUNTEER
      type: string
      description: |-
        * `CONTRACTED` - Contracted
        * `VOLUNTEER` - Volunteer
...
    MediaTypeEnum:
      enum:
      - IMAGE
      - VIDEO
      - DOCUMENT
      type: string
      description: |-
        * `IMAGE` - Image
        * `VIDEO` - Video
        * `DOCUMENT` - Document
...
    NotificationReadiness:
      type: object
      properties:
        notification_ready:
          type: boolean
        active_device_count:
          type: integer
        message:
          type: string
      required:
      - active_device_count
      - notification_ready
...
    PaginatedDiscoveryBusinessCardList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryBusinessCard'
...
    PublicBusiness:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        business_name:
          type: string
          maxLength: 255
        district:
          $ref: '#/components/schemas/DistrictEnum'
        district_label:
          type: string
          readOnly: true
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        listing_type_label:
          type: string
          readOnly: true
        is_featured:
          type: boolean
        short_description:
          type: string
          readOnly: true
        intro_text:
          type: string
          readOnly: true
        badge_text:
          type: string
          readOnly: true
        cover_image:
          type: string
          readOnly: true
        logo_image:
          type: string
          readOnly: true
        primary_marketplace_category:
          type: object
          additionalProperties: {}
          nullable: true
          readOnly: true
      required:
      - badge_text
      - business_name
      - cover_image
      - district_label
      - id
      - intro_text
      - listing_type_label
      - logo_image
      - primary_marketplace_category
      - short_description
    PublicBusinessCategoryOverview:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
      - description
      - id
      - name
    PublicBusinessDetailResponse:
      type: object
      properties:
        business:
          $ref: '#/components/schemas/PublicBusiness'
        media:
          type: array
          items:
            $ref: '#/components/schemas/PublicMediaAsset'
        active_offers:
          type: array
          items:
            $ref: '#/components/schemas/PublicOffer'
        category_overview:
          type: array
          items:
            $ref: '#/components/schemas/PublicBusinessCategoryOverview'
        server_time:
          type: string
          format: date-time
      required:
      - active_offers
      - business
      - category_overview
      - media
      - server_time
    PublicBusinessListResponse:
      type: object
      properties:
        count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/PublicBusiness'
      required:
      - count
      - results
    PublicBusinessMenuResponse:
      type: object
      properties:
        business:
          $ref: '#/components/schemas/PublicBusiness'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/PublicCategory'
        active_offers:
          type: array
          items:
            $ref: '#/components/schemas/PublicOffer'
      required:
      - active_offers
      - business
      - categories
    PublicCategory:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 120
        description:
          type: string
        menu_items:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
      required:
      - id
      - menu_items
      - name
    PublicMediaAsset:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        media_type:
          $ref: '#/components/schemas/MediaTypeEnum'
        asset_role:
          $ref: '#/components/schemas/AssetRoleEnum'
        url:
          type: string
          readOnly: true
        alt_text:
          type: string
          maxLength: 255
        sort_order:
          type: integer
      required:
      - id
      - url
    PublicOffer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 160
        short_description:
          type: string
          maxLength: 255
        description:
          type: string
        label:
          type: string
          maxLength: 64
        tag:
          type: string
          maxLength: 64
        offer_price_amount:
          type: integer
          description: Kuruş cinsinden kampanya fiyatı
        is_featured:
          type: boolean
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        is_live:
          type: boolean
          readOnly: true
        image:
          type: string
          readOnly: true
      required:
      - ends_at
      - id
      - image
      - is_live
      - offer_price_amount
      - starts_at
      - title
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...